% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_kulaBar.R
\name{add_kulaBar}
\alias{add_kulaBar}
\title{Add a custom colour bar}
\usage{
add_kulaBar(
  kula,
  range,
  increments = NULL,
  tickAxis = 1,
  nameAxis = NULL,
  labels = NULL,
  mar = NULL,
  labelFirst = 1,
  labelEvery = 1,
  ...
)
}
\arguments{
\item{kula}{The colours to plot in the colour bar. Either a vector of
colours (e.g. a colour_scheme from \code{khroma}), or a string that is fed
into \code{kulaK()} (a wrapper around \code{khroma::colour()}). Usually, it makes
more sense to provide a vector (e.g. the output of \code{kulaR::kulaK()})
because this provides more flexibility and is likely being used in
whatever the colour bar refers to.}

\item{range}{vector: min and max values for the kulaBar. The min will
always be plotted to the left of horizontal bars, or at the bottom of
vertical bar.}

\item{increments}{numeric: How much change is represented by one block of
colour along the bar? If kula is a colour scheme, this argument is
ignored and the increments are calculated automatically by dividing the
full range by the length of the kula vector. If 'kula' is a string (e.g.
"BuRd"), the increments cannot be calculated automatically so this value
is necessary; in such a case, the difference between the 'range' values
must be an exact multiple of the 'increment' argument.}

\item{tickAxis}{numeric: Which side of the axis should the ticks and labels
be added? This determines orientation: 1 (bottom) and 3 (top) are
horizontal, 2 (left) and 4 (right) are vertical. Default is 1.}

\item{nameAxis}{numeric: Which side of the axis should the name be added?}

\item{labels}{vector: If not NULL, values are used as labels for the ticks
along the axis instead of the values calculated from 'range' and
'increments'. Useful for adding something such as "<0" for example.
Labels must be provided for all ticks (i.e. assuming that 'labelEvery' =
1), even if 'labelEvery' is not 1. See examples.}

\item{mar}{vector: Set the margins around the kulaBar (as par(mar = mar)).}

\item{labelFirst}{Which is the first tick that should be labelled? Use the
index, not the value. See \code{figuR::add_axis()}.}

\item{labelEvery}{Labels should be added every how many ticks? See
\code{figuR::add_axis()}.}

\item{...}{Any additional parameters not listed above that can be fed into
figuR::add_axis().}
}
\description{
Add a colour bar / legend with more fine control. Make sure to
prepare the plot area before using this function (using e.g. \code{layout()}).
The function essentially plots a matrix with a single row / column using
the \code{graphics::image()}, and then uses \code{figuR::add_axis()}.
}
\examples{
  \dontrun{
    # Make changes to the colours beforehand
    kk <- kulaK("BuRd", 14, c(3, 7), skipMid = 1)
    add_kulaBar(kk, c(-3, 7) * 4)

    # Be careful with the range and the labels if doing unusual palettes
    kk <- kulaK(palette = "bamako", count = 14,
                reverse = TRUE,
                addFirst = "magenta", include = 1:10)
    add_kulaBar(kula = kk, range = c(1, 16),
                labels = c("", "<0", 1:10), labelEvery = 2, labelFirst = 2)

    # Without first defining the colour palette
    add_kulaBar(kula = "romaO", range = c(0, 360), increments = 45)
  }


}
