% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_kulaBar.R
\name{annotate_kulaBar}
\alias{annotate_kulaBar}
\title{Add text around a kulaBar to explain the colours}
\usage{
annotate_kulaBar(
  aboveLeft = "",
  above = "",
  aboveRight = "",
  belowLeft = "",
  below = "",
  belowRight = "",
  offset_v = 0.1,
  offset_h = 0,
  cex = 0.75
)
}
\arguments{
\item{aboveLeft, above, aboveRight}{"string": Text above the kulaBar in the
respective positions.}

\item{belowLeft, below, belowRight}{"string": Text below the kulaBar in the
respective positions.}

\item{offset_v}{numeric: How far from the kulaBar should the text be
located vertically? Positive values move text away from the kulaBar; unit
is fractions of the kulaBar height. If two values are provided, the first
values moves the text strings above the bar, and the second the strings
below; for a single value, both sets are moved the same distance \emph{away}
(if positive) or \emph{towards} (if negative) from the kulaBar.}

\item{offset_h}{numeric: As for 'offset_v' but horizontal offsets. If two
values are supplied, the first applies to the left side and the second to
the right side. Units are also fractions of the kulaBar \emph{height}.}

\item{cex}{numeric: How large should the text be?}
}
\description{
This function is particularly useful when a kulaBar displays
differences that are not immediately intuitive. The function simply adds
text at up to 6 locations around the kulaBar. It needs to be called
immediately after using \code{add_kulaBar()}. It currently only works for
horizontal kulaBars.
}
