% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_domR.R
\name{count_decimal_places}
\alias{count_decimal_places}
\title{Counts how many digits are after the decimal point}
\usage{
count_decimal_places(x)
}
\arguments{
\item{x}{numeric: The value}
}
\description{
This function is useful when we are automating things and need
to round, but aren't always sure what precision we need to round to. The
code was taken from this stackoverflow answer by daroczig:
\url{https://stackoverflow.com/questions/5173692/how-to-return-number-of-decimal-places-in-r},
and I do not entirely understand how it works.
}
