% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kulaInfo.R
\name{get_kulaInfo}
\alias{get_kulaInfo}
\title{Automate colour ranges for plotting}
\usage{
get_kulaInfo(x, palette = NULL, skipMid = NULL, tickCount = 12)
}
\arguments{
\item{x}{The data to calculate the kulaInfo for. Can be a vector, a matrix,
a list, or SpatRaster (requires the
\href{https://cran.r-project.org/web/packages/terra/terra.pdf}{terra}
package).}

\item{palette}{"string": Overwrite the default palettes ("BuRd" for
diverging values, "lajolla" for ascending sequential values, and "devon"
for descending sequential values. Use \code{list_kula_palettes()} for a list
of accepted values.}

\item{skipMid}{numeric: Should more intense colours be used? See \code{kulaK()}
for an explanation.}

\item{tickCount}{numeric: The ideal number of ticks, fed directly into the
\code{pretty()} function; it attempts to set this number of ticks, but may
choose more or less if there are prettier alternatives.}
}
\value{
a list, containing:
\itemize{
\item zKula          the colour palette
\item zRange         the minimum & maximum values
\item zIncrements    the increment between steps
\item zPositives     how many values are above 0?
\item zNegatives     how many values are below 0?
}
}
\description{
We often want to automate plots of different variables, or for
different periods etc. This functions helps with automating the colours
by returning a list of useful parts that can be fed into other functions
(e.g. \code{kulaK()}, \code{add_kulaBar()}).
}
