% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaL.R
\name{kulaL}
\alias{kulaL}
\title{Returns different shades of grey, ideal for grid lines}
\usage{
kulaL(shade = 4, opacity = 10)
}
\arguments{
\item{shade}{numeric: A value between 0 (white) and 10 (black).}

\item{opacity}{numeric: A value between 0 (invisible) and 10 (opaque).}
}
\description{
For plotting lines, we often want a simple grey or black
colour, which can take require unnecessarily verbose code; this function
simplifies all of that for us and provides a wrapper around the
\code{\link[grDevices:rgb]{grDevices::rgb()}} function that is more consistent with the "kulaR"
syntax. See also \code{\link[=kulaT]{kulaT()}}, which is the same, but swaps the order of the
arguments to make it slightly easier for changing the line transparency.
}
