% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaQ.R
\name{kulaQ}
\alias{kulaQ}
\title{Returns the "bright" qualitative khroma colour scheme}
\usage{
kulaQ(include = NULL, ...)
}
\arguments{
\item{include}{"string" or c(vector): Return only a subset of the colours.
Can either be defined by their name, or their index.}

\item{...}{Any arguments that can be used by the \code{kulaK()} function,
although "skipMid", and "reverse" don't make much sense.}
}
\description{
Just a wrapper around \code{kulaK()} (which is a wrapper around
khroma) for when we want to use a qualitative colour scheme. Either
return the full "bright" palette (the default), or add an index /
character strings to select a subset of the colours (useful for e.g.
nicer line colours and matching lines across plots). The options are:
"red", "blue", "green", "yellow", "cyan", "purple" and "grey".
}
\examples{
  plot(1:7, 1:7)
  abline(h = 1:7, col = kulaQ(), lwd = 5)
  abline(v = 1:7, col = kulaQ("red"), lwd = 5)
  plot(1:7, 1:7)
  abline(v = 1:7, col = kulaQ("red", "blue"), lwd = 5)
  abline(h = 1:7, col = kulaQ(1:3), lwd = 5)

}
