% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaQ.R
\name{kulaQ2}
\alias{kulaQ2}
\title{Returns the "vibrant" qualitative khroma colour scheme}
\usage{
kulaQ2(include = NULL, ...)
}
\arguments{
\item{include}{"string" or c(vector): Return only a subset of the colours.
Can either be defined by their name, or their index.}

\item{...}{Any arguments that can be used by the \code{\link[=kulaK]{kulaK()}} function,
although "skipMid", and "reverse" don't make much sense.}
}
\description{
Just a wrapper around \code{\link[=kulaK]{kulaK()}} (which is a wrapper around
\code{\link[khroma:colour]{khroma::colour()}}) for when we want to use a qualitative colour scheme.
Either return the full "vibrant" palette (the default), or add an index /
character strings to select a subset of the colours (useful for e.g.
nicer line colours and matching lines across plots). The options are:
"orange", "blue", "cyan", "magenta", "red", "teal", and "grey".
}
\examples{
  plot(1:7, 1:7)
  abline(h = 1:7, col = kulaQ2(), lwd = 5)
  abline(v = 1:7, col = kulaQ2("orange"), lwd = 5)
  plot(1:7, 1:7)
  abline(v = 1:7, col = kulaQ2("orange", "cyan"), lwd = 5)
  abline(h = 1:7, col = kulaQ2(1:3), lwd = 5)

}
