% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaQ.R
\name{kulaQ3}
\alias{kulaQ3}
\title{Returns the "highcontrast" qualitative khroma colour scheme}
\usage{
kulaQ3(include = NULL, ...)
}
\arguments{
\item{include}{"string" or c(vector): Return only a subset of the colours.
Can either be defined by their name, or their index.}

\item{...}{Any arguments that can be used by the \code{\link[=kulaK]{kulaK()}} function,
although "skipMid", and "reverse" don't make much sense.}
}
\description{
Just a wrapper around \code{\link[=kulaK]{kulaK()}} (which is a wrapper around
\code{\link[khroma:colour]{khroma::colour()}}) for when we want to use a qualitative colour scheme.
Either return the full "highcontrast" palette (the default), or add an
index / character strings to select a subset of the colours (useful for
e.g. nicer line colours and matching lines across plots). The options
are: "blue", "yellow" and "red".
}
\examples{
  plot(1:3, 1:3)
  abline(h = 1:3, col = kulaQ3(), lwd = 5)
  abline(v = 1:3, col = kulaQ3("blue"), lwd = 5)
  plot(1:3, 1:3)
  abline(v = 1:3, col = kulaQ3("yellow", "blue"), lwd = 5)

}
