% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaL.R
\name{kulaT}
\alias{kulaT}
\title{Returns semi-transparent grey colours, ideal for grid lines}
\usage{
kulaT(opacity = 4, shade = 10)
}
\arguments{
\item{opacity}{numeric: A value between 0 (invisible) and 10 (opaque).}

\item{shade}{numeric: A value between 0 (white) and 10 (black).}
}
\description{
For plotting lines, we often want a simple grey or black
colour, which can take require unnecessarily verbose code; this function
simplifies all of that for us and provides a wrapper around the \code{rgb()}
function that is more consistent with the "kulaR" syntax. See also
\code{kulaL()}, which is the same, but swaps the order of the arguments to
make it slightly easier for changing the line colour.
}
