% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kula_palettes.R
\name{list_kula_palettes}
\alias{list_kula_palettes}
\title{Access the possible khroma colour schemes}
\usage{
list_kula_palettes(type = NULL, print = TRUE)
}
\arguments{
\item{type}{"string": Which kind of colour scheme are you interested in?
Options are "diverging", "sequential", and "qualitative".}

\item{print}{BINARY: Should the kula palettes be printed? TRUE by default
(as it is the point of this function!), but FALSE suppresses so it can be
used in other functions.}
}
\description{
A wrapper around khroma::info() that is fed into the
other "list_kulaX_palettes" functions. kulaInfo is invisibly returned.
}
