% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kula_palettes.R
\name{list_kulaC_palettes}
\alias{list_kulaC_palettes}
\title{Print out the possible circular khroma colour schemes}
\usage{
list_kulaC_palettes(print = TRUE)
}
\arguments{
\item{print}{BINARY: Should the kula palettes be printed? TRUE by default
(as it is the point of this function!), but FALSE suppresses so it can be
used in other functions.}
}
\description{
Just a way of using khroma::info(), but only showing
circular colour schemes. kulaInfo is invisibly returned.
}
