% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kula_palettes.R
\name{list_kulaD_palettes}
\alias{list_kulaD_palettes}
\title{Print out the possible diverging khroma colour schemes}
\usage{
list_kulaD_palettes(print = TRUE)
}
\arguments{
\item{print}{BINARY: Should the kula palettes be printed? TRUE by default
(as it is the point of this function!), but FALSE suppresses so it can be
used in other functions.}
}
\description{
A wrapper around \code{khroma::info()}, but only showing diverging
colour schemes. kulaInfo is invisibly returned.
}
