% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kula_palettes.R
\name{list_kulaQ_palettes}
\alias{list_kulaQ_palettes}
\title{Print out the possible qualitative khroma colour schemes}
\usage{
list_kulaQ_palettes(print = TRUE)
}
\arguments{
\item{print}{BINARY: Should the kula palettes be printed? TRUE by default
(as it is the point of this function!), but FALSE suppresses so it can be
used in other functions.}
}
\description{
A wrapper around \code{\link[khroma:info]{khroma::info()}}, but only showing
qualitative colour schemes. kulaInfo is invisibly returned.
}
