% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_kula_palettes.R
\name{list_kulaS_palettes}
\alias{list_kulaS_palettes}
\title{Print out the possible sequential khroma colour schemes}
\usage{
list_kulaS_palettes(print = TRUE)
}
\arguments{
\item{print}{BINARY: Should the kula palettes be printed? TRUE by default
(as it is the point of this function!), but FALSE suppresses so it can be
used in other functions.}
}
\description{
A wrapper around \code{khroma::info()}, but only showing sequential
colour schemes. kulaInfo is invisibly returned. Use \code{list_kulaC_palettes}
for circular palettes.
}
