% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kula_palettes.R
\name{plot_kula_palettes}
\alias{plot_kula_palettes}
\title{Display the khroma colour palettes}
\usage{
plot_kula_palettes(orientation = "v", type = NULL)
}
\arguments{
\item{orientation}{Should the colour palettes be displated vertically ("v")
or horizontally ("h")?}

\item{type}{"string": Which kind of colour scheme are you interested in?
Options are "diverging", "sequential", and "qualitative".}
}
\description{
Plot all the possible colour palettes from khroma
}
