% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kula_palettes.R
\name{plot_kulaC_palettes}
\alias{plot_kulaC_palettes}
\title{Display the kulaC (circular) colour palettes}
\usage{
plot_kulaC_palettes()
}
\description{
Plot all the possible kulaC palettes to compare & select from.
}
