% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kula_palettes.R
\name{plot_kulaQ_palettes}
\alias{plot_kulaQ_palettes}
\title{Display the kulaQ (qualitative) colour palettes}
\usage{
plot_kulaQ_palettes(orientation = "v")
}
\arguments{
\item{orientation}{Should the colour palettes be displated vertically ("v")
or horizontally ("h")?}
}
\description{
Plot all the possible kulaQ palettes to compare & select from.
}
