% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kula_palettes.R
\name{plot_kulaS_palettes}
\alias{plot_kulaS_palettes}
\title{Display the kulaS (sequential) colour palettes}
\usage{
plot_kulaS_palettes(orientation = "v")
}
\arguments{
\item{orientation}{Should the colour palettes be displated vertically ("v")
or horizontally ("h")?}
}
\description{
Plot all the possible kulaS palettes to compare & select from.
}
