% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_if_null}
\alias{set_if_null}
\title{Sets a default value for a NULL argument}
\usage{
set_if_null(x, defaultValue)
}
\arguments{
\item{x}{Which argument is being checked?}

\item{defaultValue}{What's the value if the argument is NULL?}
}
\description{
In some functions, it is necessary to create an argument that
has a default value defined by the other arguments (i.e. calculated
within the function), but that can nevertheless be overwritten if
required. We therefore set the argument value as NULL, and can use this
function within another function to determine whether the argument's
value needs to be calculated or not. This function just makes it cleaner,
clearer and quicker to see what is happening compared to the default
syntax for this process.
}
