% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_kulaBar.R
\name{annotate_kulaBar}
\alias{annotate_kulaBar}
\title{Add text around a kulaBar to explain the colours}
\usage{
annotate_kulaBar(
  aboveLeft = "",
  above = "",
  aboveRight = "",
  midLeft = "",
  mid = "",
  midRight = "",
  belowLeft = "",
  below = "",
  belowRight = "",
  offset_v = c(0.1, 0.15),
  offset_h = 0,
  cex = 0.75,
  col = "black"
)
}
\arguments{
\item{aboveLeft, above, aboveRight}{"string": Text above the kulaBar in the
respective positions.}

\item{midLeft, mid, midRight}{"string": Text aligned centred against the
kulaBar in the respective positions.}

\item{belowLeft, below, belowRight}{"string": Text below the kulaBar in the
respective positions.}

\item{offset_v}{numeric: How far from the kulaBar should the text be
located vertically?

Positive values move text away from the kulaBar, and the unit is in
fractions of the kulaBar height. If two values are provided, the first
values moves the text strings that are above the bar, and the second the
strings that are below; for a single value, both sets are moved the same
distance \emph{away} (if positive) or \emph{towards} (if negative) from the
kulaBar.

'midLeft', 'mid' and 'midRight' text cannot be vertically offset, they
just stick to the centre of the kulaBar.}

\item{offset_h}{numeric: As for 'offset_v' but horizontal offsets. If two
values are supplied, the first applies to the left side and the second to
the right side. Units are fractions of the kulaBar \emph{width}.

'mid', 'above' and 'below' text cannot be horizontally offset, they just
stick to the centre of the kulaBar.}

\item{cex}{numeric: How large should the text be?}

\item{col}{What colour should the text be?}
}
\description{
This function is particularly useful when a kulaBar displays
differences that are not immediately intuitive. The function simply adds
text at up to 9 locations around the kulaBar. It needs to be called
immediately after using \code{add_kulaBar()}.
}
\examples{
  \dontrun{
    # horizontal bar
    kk <- kulaD(10)
    add_kulaBar(kk, c(-5, 5), mar = c(5, 4, 5, 4))
    annotate_kulaBar("colder",    NA,   "warmer",
                     "OR",        NULL, "OR",
                     "less warm", "",   "less cold")
    # vertical bar
    kk <- kulaD(10)
    add_kulaBar(kk, c(-5, 5), mar = c(5, 15, 5, 15), tickAxis = 2)
    annotate_kulaBar("warmer", "OR", "less cold",
                     NA,       NULL, "",
                     "colder", "OR", "less warm")
  }

}
