% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kulaInfo.R
\name{get_kulaInfo}
\alias{get_kulaInfo}
\title{Automate colour ranges for plotting}
\usage{
get_kulaInfo(
  x,
  palette = NULL,
  skipMid = NULL,
  reverse = FALSE,
  tickCount = 12
)
}
\arguments{
\item{x}{The data to calculate the kulaInfo for. Can be a vector, a matrix,
a list, or SpatRaster (see \code{\link[terra:rast]{terra::rast()}}).}

\item{palette}{"string": Overwrite the default palettes ("BuRd" for
diverging values, "lajolla" for ascending sequential values, and "devon"
for descending sequential values. Use \code{\link[=list_kula_palettes]{list_kula_palettes()}} for a list
of accepted values.}

\item{skipMid}{numeric: Should more intense colours be used in a diverging
colour scheme? See \code{\link[=kulaK]{kulaK()}} for an explanation.}

\item{reverse}{logical: Should the colour palette be reversed?}

\item{tickCount}{numeric: The ideal number of ticks, fed directly into the
\code{\link[base:pretty]{base::pretty()}} function; it attempts to set this number of ticks, but
may choose more or less if there are prettier alternatives.}
}
\value{
a list, containing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* zKulas         the colours
* zRange         the minimum & maximum values
* zIncrement     the increment between steps
* zPositives     how many values are above 0?
* zNegatives     how many values are below 0?
}\if{html}{\out{</div>}}
}
\description{
We often want to automate plots of different variables, or for
different periods etc. This functions helps with automating the colours
by returning a list of useful parts that can be fed into other functions
(e.g. \code{\link[=kulaK]{kulaK()}}, \code{\link[=add_kulaBar]{add_kulaBar()}}).
}
