% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaC.R
\name{kulaC2}
\alias{kulaC2}
\title{Returns a circular blue-red khroma colour scheme}
\usage{
kulaC2(count, ...)
}
\arguments{
\item{count}{numeric: How many colours should be returned?}

\item{...}{Any arguments that can be used by the \code{kulaK()} function.}
}
\description{
Just a wrapper around \code{kulaK()} (which is a wrapper around
khroma) for when we want to use the "vikO" colour scheme.
}
\examples{
  mm <- matrix(sample(1:100, 64), nrow = 8)
  image(mm, col = kulaC2(100))

}
