% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaD.R
\name{kulaD}
\alias{kulaD}
\title{Returns a diverging blue-red khroma colour scheme}
\usage{
kulaD(count, ...)
}
\arguments{
\item{count}{numeric: How many colours should be returned?}

\item{...}{Any arguments that can be used by the \code{\link[=kulaK]{kulaK()}} function.}
}
\description{
Just a wrapper around \code{\link[=kulaK]{kulaK()}} (which is a wrapper around
\code{\link[khroma:colour]{khroma::colour()}}) for when we want to use the "BuRd" colour scheme.
}
\examples{
  mm <- matrix(sample(1:100, 64), nrow = 8)
  image(mm, col = kulaD(10))

}
