% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaD.R
\name{kulaD2}
\alias{kulaD2}
\title{Returns a diverging purple-green khroma colour scheme}
\usage{
kulaD2(count, ...)
}
\arguments{
\item{count}{numeric: How many colours should be returned?}

\item{...}{Any arguments that can be used by the \code{kulaK()} function.}
}
\description{
Just a wrapper around \code{kulaK()} (which is a wrapper around
khroma) for when we want to use the "PRGn" colour scheme.
}
\examples{
  mm <- matrix(sample(1:100, 64), nrow = 8)
  image(mm, col = kulaD2(10))

}
