% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kulaS.R
\name{kulaS2}
\alias{kulaS2}
\title{Returns a green-through-yellow sequential khroma colour scheme}
\usage{
kulaS2(count, ...)
}
\arguments{
\item{count}{numeric: How many colours should be returned?}

\item{...}{Any arguments that can be used by the \code{\link[=kulaK]{kulaK()}} function.}
}
\description{
Just a wrapper around \code{\link[=kulaK]{kulaK()}} (which is a wrapper around
\code{\link[khroma:colour]{khroma::colour()}}) for when we want to use the "bamako" colour scheme.
}
\examples{
  mm <- matrix(sample(1:100, 64), nrow = 8)
  image(mm, col = kulaS2(10))

}
