% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kula_palettes.R
\name{plot_kulaD_palettes}
\alias{plot_kulaD_palettes}
\title{Display the kulaD (diverging) colour palettes}
\usage{
plot_kulaD_palettes(orientation = "v")
}
\arguments{
\item{orientation}{Should the colour palettes be displated vertically ("v")
or horizontally ("h")?}
}
\description{
Plot all the possible kulaD palettes to compare & select from.
}
